+++ tango/common/dataset_dict.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass, field
-from typing import Mapping, Any, Sequence, TypeVar, Generic, Optional
+from typing import Mapping, Any, Sequence, TypeVar, Generic, Optional, Iterator
 
 from .det_hash import CustomDetHash
 
@@ -8,10 +8,10 @@ T = TypeVar("T")
 
 
 @dataclass
-class DatasetDict(CustomDetHash, Generic[T]):
+class DatasetDict(CustomDetHash, Generic[T], Mapping[str, Sequence[T]]):
     """
-    This definition of a dataset combines all splits and arbitrary metadata into
-    one handy class.
+    A generic :class:`~collections.abc.Mapping` class of split names (:class:`str`) to datasets
+    (``Sequence[T]``).
     """
 
     splits: Mapping[str, Sequence[T]]
@@ -28,13 +28,18 @@ class DatasetDict(CustomDetHash, Generic[T]):
     """
     A unique fingerprint associated with the data.
 
-    When this is specified it will be used by the step caching mechanism to
-    determine when the data has changed. Otherwise the caching mechanism will
-    have to fall back to pickling the whole dataset dict to calculate a hash,
-    which can be slow, so it's recommend you set this whenever possible.
+    .. important::
+        When this is specified it will be used by the step caching mechanism to
+        determine when the data has changed. Otherwise the caching mechanism will
+        have to fall back to pickling the whole dataset dict to calculate a hash,
+        which can be slow, so it's recommend you set this whenever possible.
     """
 
     def det_hash_object(self) -> Any:
+        """
+        Overrides :meth:`~tango.common.det_hash.CustomDetHash.det_hash_object` to return
+        :attr:`fingerprint`  when specified instead of ``self``.
+        """
         if self.fingerprint is not None:
             return self.fingerprint
         else:
@@ -42,15 +47,30 @@ class DatasetDict(CustomDetHash, Generic[T]):
 
     def __getitem__(self, split: str) -> Sequence[T]:
         """
-        Get a split.
+        Get a split in :attr:`splits`.
         """
         return self.splits[split]
 
+    def __contains__(self, split: str) -> bool:  # type: ignore[override]
+        """
+        Checks if :attr:`splits` contains the given split.
+        """
+        return split in self.splits
+
+    def __iter__(self) -> Iterator[str]:
+        """
+        Returns an iterator over the keys in :attr:`splits`.
+        """
+        return iter(self.splits.keys())
+
     def __len__(self) -> int:
         """
-        Returns the number of splits.
+        Returns the number of splits in :attr:`splits`.
         """
         return len(self.splits)
 
     def keys(self):
+        """
+        Returns the split names in :attr:`splits`.
+        """
         return self.splits.keys()